+++ CHANGELOG.md
@@ -7,3 +7,6 @@
 - Switches to relative lines when invoking the command
 - Highlights the cursor destination
 - Handle multiple motions in a single execution (e.g. `10j2k`)
+- Handle line start/end motions (`$`, `^`)
+- Handle word boundary motions (`w`, `e`, `b`) (supports repetitions, e.g. `10w`)
+- Handle find next/previous character motions (`F`, `f`) (supports repetitions, e.g. `2f"`)